[REFACTOR] reorganize package imports
