+++ .gitignore
@@ -4,3 +4,4 @@ weight/
 data/
 notebooks/
 checkpoint/
+.idea/
