+++ __init__.py
@@ -1,0 +1,5 @@
+import sys
+import os
+sys.path.append(os.path.join(os.path.dirname(__file__), 'external/lightweight-human-pose-estimation.pytorch'))
+from .model import *
+from .utils import *
