+++ model/__init__.py
@@ -1,0 +1,2 @@
+from .Pose3D import *
+from .GCN import *
