+++ scripts/__init__.py
@@ -1,0 +1,4 @@
+# scripts/__init__.py
+import sys
+import os
+sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..', 'external/lightweight-human-pose-estimation.pytorch'))
