+++ scripts/create_baseline_sequence.py
@@ -4,8 +4,6 @@ import pandas as pd
 from fastdtw import fastdtw
 from scipy.interpolate import interp1d
 
-import sys
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 from config import DATA_DIR
