+++ scripts/visualize_2d_skeleton_from_json.py
@@ -5,9 +5,7 @@ import numpy as np
 from tqdm import tqdm
 
 import sys
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 from config import BASE_DIR, WEIGHT_PATH, DATA_DIR
-sys.path.append(os.path.join(BASE_DIR, 'external/lightweight-human-pose-estimation.pytorch'))
 from modules.pose import Pose
