+++ utils/sequences.py
@@ -3,7 +3,7 @@ from scipy.interpolate import interp1d
 from fastdtw import fastdtw
 import os
 import sys
-sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
 from config import DATA_DIR
